[ Misc ] Clean Up `CompressedTensorsW8A8` (#6113)
